💫 UPDATE: VIP_Integration_Test to include tests of IntegrationVipConfig methods which are moved
