+++ tests/integrations/test-integration.php
@@ -9,12 +9,13 @@ namespace Automattic\VIP\Integrations;
 
 // phpcs:disable Squiz.Commenting.ClassComment.Missing, Squiz.Commenting.FunctionComment.Missing, Squiz.Commenting.FunctionComment.MissingParamComment
 
-use Automattic\VIP\Integrations\IntegrationVipConfig;
 use Env_Integration_Status;
 use ErrorException;
+use Org_Integration_Status;
 use PHPUnit\Framework\MockObject\MockObject;
 use WP_UnitTestCase;
 
+use function Automattic\Test\Utils\get_class_method_as_public;
 use function Automattic\Test\Utils\reset_custom_error_reporting;
 use function Automattic\Test\Utils\setup_custom_error_reporting;
 
@@ -67,7 +68,7 @@ class VIP_Integration_Test extends WP_UnitTestCase {
 		$this->assertEquals( [ 'config_test' ], $integration->get_config() );
 	}
 
-	public function test__calling_activate_when_the_integration_is_already_loaded_does_not_activate_the_integration_again(): void {
+	public function test__activate_does_not_activate_the_integration_again_when_the_integration_is_already_loaded(): void {
 		$this->expectException( ErrorException::class );
 		$this->expectExceptionCode( E_USER_WARNING );
 		$this->expectExceptionMessage( 'Prevented activating of integration with slug "fake" because it is already loaded.' );
@@ -84,7 +85,7 @@ class VIP_Integration_Test extends WP_UnitTestCase {
 		$this->assertFalse( $integration_mock->is_active() );
 	}
 
-	public function test__calling_activate_twice_on_same_integration_does_not_activate_the_plugin_second_time(): void {
+	public function test__activate_throws_error_when_same_integration_is_activated_twice(): void {
 		$this->expectException( ErrorException::class );
 		$this->expectExceptionCode( E_USER_WARNING );
 		$this->expectExceptionMessage( 'VIP Integration with slug "fake" is already activated.' );
@@ -137,5 +138,333 @@ class VIP_Integration_Test extends WP_UnitTestCase {
 		$integration = new FakeIntegration( 'fake' );
 
 		$this->assertFalse( $integration->is_active() );
+	}
+
+	public function test__is_active_via_vip_returns_false_if_empty_config_is_provided(): void {
+		$this->do_test_is_active_via_vip( [], false );
+	}
+
+	public function test__is_active_via_vip_returns_false_if_organization_status_is_blocked(): void {
+		$this->do_test_is_active_via_vip( [ [ 'org' => [ 'status' => Org_Integration_Status::BLOCKED ] ] ], false );
+	}
+
+	public function test__is_active_via_vip_returns_false_if_environment_status_is_blocked(): void {
+		$this->do_test_is_active_via_vip( [ [ 'env' => [ 'status' => Org_Integration_Status::BLOCKED ] ] ], false );
+	}
+
+	public function test__is_active_via_vip_returns_false_if_integration_is_blocked_on_current_network_site(): void {
+		if ( ! is_multisite() ) {
+			$this->markTestSkipped( 'Only valid for multisite.' );
+		}
+
+		$this->do_test_is_active_via_vip(
+			[
+				[
+					'env'           => [ 'status' => Env_Integration_Status::ENABLED ],
+					'network_sites' => [ '1' => [ 'status' => Env_Integration_Status::BLOCKED ] ],
+				],
+			],
+			false,
+		);
+	}
+
+	public function test__is_active_via_vip_returns_false_if_integration_is_disabled_on_current_network_site(): void {
+		if ( ! is_multisite() ) {
+			$this->markTestSkipped( 'Only valid for multisite.' );
+		}
+
+		$this->do_test_is_active_via_vip(
+			[
+				[
+					'env'           => [ 'status' => Env_Integration_Status::ENABLED ],
+					'network_sites' => [ '1' => [ 'status' => Env_Integration_Status::DISABLED ] ],
+				],
+			],
+			false,
+		);
+	}
+
+	public function test__is_active_via_vip_returns_true_if_integration_is_enabled_on_current_network_site(): void {
+		if ( ! is_multisite() ) {
+			$this->markTestSkipped( 'Only valid for multisite.' );
+		}
+
+		$this->do_test_is_active_via_vip(
+			[
+				[
+					'env'           => [ 'status' => Env_Integration_Status::DISABLED ],
+					'network_sites' => [ '1' => [ 'status' => Env_Integration_Status::ENABLED ] ],
+				],
+			],
+			true,
+		);
+	}
+
+	public function test__is_active_via_vip_returns_true_if_integration_is_not_present_on_current_network_site_but_enabled_on_environment(): void {
+		if ( ! is_multisite() ) {
+			$this->markTestSkipped( 'Only valid for multisite.' );
+		}
+
+		$this->do_test_is_active_via_vip(
+			[
+				[
+					'env' => [ 'status' => Env_Integration_Status::ENABLED ],
+				],
+			],
+			true,
+		);
+	}
+
+	public function test__is_active_via_vip_returns_false_if_integration_is_not_provided_on_current_network_site(): void {
+		if ( ! is_multisite() ) {
+			$this->markTestSkipped( 'Only valid for multisite.' );
+		}
+
+		$this->do_test_is_active_via_vip( [
+			[
+				'network_sites' => [
+					'2' => [
+						'status' => Env_Integration_Status::ENABLED,
+						'config' => [ 'site config' ],
+					],
+				],
+			],
+		], false );
+	}
+
+	public function test__is_active_via_vip_returns_false_if_integration_is_disabled_on_environment(): void {
+		$this->do_test_is_active_via_vip( [ 'env' => [ 'status' => Env_Integration_Status::DISABLED ] ], false );
+	}
+
+	public function test__is_active_via_vip_returns_true_if_integration_is_enabled_on_environment(): void {
+		$this->do_test_is_active_via_vip( [
+			[
+				'env' => [
+					'status' => Env_Integration_Status::ENABLED,
+				],
+			],
+		], true );
+	}
+
+	public function test__is_active_via_vip_returns_false_if_integration_is_not_provided_on_environment(): void {
+		$this->do_test_is_active_via_vip( [
+			[
+				'org'           => [
+					'status' => Env_Integration_Status::ENABLED,
+				],
+				'network_sites' => [
+					'2' => [
+						'status' => Env_Integration_Status::ENABLED,
+						'config' => [ 'site config' ],
+					],
+				],
+			],
+		], false );
+	}
+
+	/**
+	 * Helper function for testing `is_active_via_vip`.
+	 *
+	 * @param array   $vip_configs
+	 * @param boolean $expected_is_active_via_vip
+	 *
+	 * @return void
+	 */
+	private function do_test_is_active_via_vip(
+		$vip_configs,
+		bool $expected_is_active_via_vip
+	) {
+		$integration = $this->get_integration_with_configs( $vip_configs );
+
+		$this->assertEquals( $expected_is_active_via_vip, $integration->is_active_via_vip() );
+	}
+
+
+	public function test__get_site_configs_returns_value_from_environment_config(): void {
+		if ( is_multisite() ) {
+			$this->markTestSkipped( 'Only valid for non multisite.' );
+		}
+
+		$this->do_test_get_site_config(
+			[
+				[
+					'env'           => [
+						'status' => Env_Integration_Status::ENABLED,
+						'config' => array( 'env-config' ),
+					],
+					'network_sites' => [
+						'1' => [
+							'status' => Env_Integration_Status::ENABLED,
+							'config' => array( 'network-site-config' ),
+						],
+					],
+				],
+			],
+			array( 'env-config' ),
+		);
+	}
+
+	public function test__get_site_configs_returns_value_from_network_site_config(): void {
+		if ( ! is_multisite() ) {
+			$this->markTestSkipped( 'Only valid for multisite.' );
+		}
+
+		$this->do_test_get_site_config(
+			[
+				[
+					'env'           => [
+						'status' => Env_Integration_Status::ENABLED,
+						'config' => array( 'env-config' ),
+					],
+					'network_sites' => [
+						'1' => [
+							'status' => Env_Integration_Status::ENABLED,
+							'config' => array( 'network-site-config' ),
+						],
+					],
+				],
+			],
+			array( 'network-site-config' ),
+		);
+	}
+
+	public function test__get_site_configs_with_cascading_config(): void {
+		if ( ! is_multisite() ) {
+			$this->markTestSkipped( 'Only valid for multisite.' );
+		}
+
+		$this->do_test_get_site_config(
+			[
+				[
+					'env'           => [
+						'status'         => Env_Integration_Status::ENABLED,
+						'config'         => array( 'env-config' ),
+						'cascade_config' => true,
+					],
+					'network_sites' => [
+						'1' => [
+							'status' => Env_Integration_Status::ENABLED,
+						],
+					],
+				],
+			],
+			array( 'env-config' ),
+		);
+	}
+
+	/**
+	 * Helper function for testing `get_site_configs`.
+	 *
+	 * @param array $vip_configs
+	 * @param mixed $expected_get_site_config
+	 *
+	 * @return void
+	 */
+	private function do_test_get_site_config(
+		$vip_configs,
+		$expected_get_site_config
+	) {
+		$mock = $this->get_integration_with_configs( $vip_configs );
+
+		$this->assertEquals( $expected_get_site_config, $mock->get_site_configs() );
+	}
+
+	public function test__get_value_from_vip_config_trigger_error_if_invalid_argument_is_passed(): void {
+		$this->expectException( ErrorException::class );
+		$this->expectExceptionCode( E_USER_WARNING );
+		$this->expectExceptionMessage( 'config_type param (invalid) must be one of org, env or network_sites.' );
+		$mocked_vip_config = [];
+
+		$this->do_test_get_value_from_config( $mocked_vip_config, 'invalid', 'key', '' );
+	}
+
+	public function test__get_value_from_vip_config_returns_null_if_given_config_type_have_no_data(): void {
+		$mocked_vip_config = [];
+
+		$this->do_test_get_value_from_config( $mocked_vip_config, 'org', 'status', null );
+	}
+
+	public function test__get_value_from_vip_config_returns_config_from_organization_data(): void {
+		$mocked_vip_config = [
+			'org' => [
+				'status' => Org_Integration_Status::BLOCKED,
+				'config' => array( 'client_configs' ),
+			],
+		];
+
+		$this->do_test_get_value_from_config( $mocked_vip_config, 'org', 'status', Org_Integration_Status::BLOCKED );
+		$this->do_test_get_value_from_config( $mocked_vip_config, 'org', 'config', array( 'client_configs' ) );
+	}
+
+	public function test__get_value_from_vip_config_returns_config_of_current_network_site(): void {
+		if ( ! is_multisite() ) {
+			$this->markTestSkipped( 'Only valid for multisite.' );
+		}
+
+		$mocked_vip_config = [
+			'env'           => [
+				'status' => Env_Integration_Status::BLOCKED,
+				'config' => array( 'env_configs' ),
+			],
+			'network_sites' => [
+				'1' => [
+					'status' => Env_Integration_Status::ENABLED,
+					'config' => array( 'network_site_1_configs' ),
+				],
+				'2' => [
+					'status' => Env_Integration_Status::ENABLED,
+					'config' => array( 'network_site_2_configs' ),
+				],
+			],
+		];
+
+		$this->do_test_get_value_from_config( $mocked_vip_config, 'network_sites', 'status', Env_Integration_Status::ENABLED );
+		$this->do_test_get_value_from_config( $mocked_vip_config, 'network_sites', 'config', array( 'network_site_1_configs' ) );
+	}
+
+	public function test__get_value_from_vip_config_returns_null_if_non_existent_key_is_passed(): void {
+		$mocked_vip_config = [
+			'env' => [
+				'status' => Env_Integration_Status::BLOCKED,
+				'config' => array( 'env_configs' ),
+			],
+		];
+
+		$this->do_test_get_value_from_config( $mocked_vip_config, 'env', 'invalid_key', null );
+	}
+
+	/**
+	 * Helper function for testing `get_value_from_vip_config`.
+	 *
+	 * @param array      $vip_configs
+	 * @param string     $config_type
+	 * @param string     $key
+	 * @param null|array $expected_value_from_vip_config
+	 *
+	 * @return void
+	 */
+	private function do_test_get_value_from_config(
+		array $vip_configs,
+		string $config_type,
+		string $key,
+		$expected_value_from_vip_config
+	): void {
+		$integration  = $this->get_integration_with_configs( $vip_configs );
+		$config_value = get_class_method_as_public( Integration::class, 'get_value_from_config' )->invoke( $integration, $vip_configs, $config_type, $key );
+
+		$this->assertEquals( $expected_value_from_vip_config, $config_value );
+	}
+
+	/**
+	 * Get `Integration` with configs.
+	 *
+	 * @param array $vip_configs
+	 */
+	private function get_integration_with_configs( $vip_configs ): Integration {
+		$integration = new FakeIntegration( 'fake' );
+
+		$integration->set_vip_configs( $vip_configs );
+
+		return $integration;
 	}
 }
